+++ cmd/buf/internal/command/beta/registry/webhook/webhooklist/webhooklist.go
@@ -16,7 +16,6 @@ package webhooklist
 
 import (
 	"context"
-	"encoding/json"
 
 	"buf.build/go/app/appcmd"
 	"buf.build/go/app/appext"
@@ -25,6 +24,7 @@ import (
 	"github.com/bufbuild/buf/private/gen/proto/connect/buf/alpha/registry/v1alpha1/registryv1alpha1connect"
 	registryv1alpha1 "github.com/bufbuild/buf/private/gen/proto/go/buf/alpha/registry/v1alpha1"
 	"github.com/bufbuild/buf/private/pkg/connectclient"
+	"github.com/bufbuild/buf/private/pkg/protoencoding"
 	"github.com/spf13/pflag"
 )
 
@@ -117,7 +117,7 @@ func run(
 		_, _ = container.Stdout().Write([]byte("[]"))
 		return nil
 	}
-	webhooksJSON, err := json.MarshalIndent(resp.Msg.GetWebhooks(), "", "\t")
+	webhooksJSON, err := protoencoding.NewJSONMarshaler(nil, protoencoding.JSONMarshalerWithIndent()).Marshal(resp.Msg)
 	if err != nil {
 		return err
 	}